add score and gold reward to unit table

+++ app/game/network/server/PlayerStatsInitPacket.java
@@ -13,7 +13,7 @@ public class PlayerStatsInitPacket extends PlayerStatsUpdatePacket {
   protected Integer goldPerTick;
 
   public PlayerStatsInitPacket(Long lives, Long gold, Integer goldPerTick) {
-    super(0L, lives, gold, 0L, lives, gold);
+    super(0L, lives, gold, 0, lives.intValue(), gold.intValue());
     this.type = PacketType.PlayerStatsInitPacket;
     this.goldPerTick = goldPerTick;
   }
